+++ packages/ub-pub/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+- `LocalRepository.clone()` will copy a private `_localData` of source repository to destination (as reference!)
 
 ## [5.7.1] - 2021-01-19
 ## [5.7.0] - 2021-01-17